Adds List to internalversions types
This is used in order to get apimachinery to work with Kedge in order to
support OpenShift.

Since go-client specifically uses a version of apimachinery as well as
the fact that the main repo (kubernetes/kubernetes) uses a different
version, adding this "hack" fixes the compatibility issue in order to
get Kedge to work in reading OpenShift structs and types.